requirements.txt
fastapi
uvicorn
python-multipart
pyjwt
bcrypt
passlib
google-generativeai
Pillow
sqlalchemy
pydantic
python-dotenv
